+++ src/app/static/src/app/hintVersion.ts
@@ -1,1 +1,1 @@
-export const currentHintVersion = "1.39.1";
+export const currentHintVersion = "1.41.1";