+++ src/app/static/src/app/store/plottingSelections/mutations.ts
@@ -11,6 +11,7 @@ import {DataType} from "../surveyAndProgram/surveyAndProgram";
 type PlottingSelectionsMutation = Mutation<PlottingSelectionsState>
 
 export interface PlottingSelectionsMutations {
+    updateCalibratePlotSelections: PlottingSelectionsMutation,
     updateBarchartSelections: PlottingSelectionsMutation,
     updateBubblePlotSelections: PlottingSelectionsMutation,
     updateSAPChoroplethSelections: PlottingSelectionsMutation,
@@ -21,6 +22,9 @@ export interface PlottingSelectionsMutations {
 }
 
 export const mutations: MutationTree<PlottingSelectionsState> & PlottingSelectionsMutations = {
+    updateCalibratePlotSelections(state: PlottingSelectionsState, action: PayloadWithType<Partial<BarchartSelections>>) {
+        state.calibratePlot = {...state.calibratePlot, ...action.payload};
+    },
     updateBarchartSelections(state: PlottingSelectionsState, action: PayloadWithType<Partial<BarchartSelections>>) {
         state.barchart = {...state.barchart, ...action.payload};
     },